api/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.1
PyJWT>=2.8.0
psycopg2-binary>=2.9.7
python-dotenv>=1.0.0
pydantic>=2.4.2